Allowing MIN Integer Value

+++ Compiler/Lexer/Tokens/IntegerToken.cs
@@ -1,8 +1,8 @@
 ﻿namespace RappiSharp.Compiler.Lexer.Tokens {
   internal class IntegerToken : Token {
-    public int Value { get; }
+    public long Value { get; }
 
-    public IntegerToken(Location? location, int value) : 
+    public IntegerToken(Location? location, long value) : 
       base(location) {
       Value = value;
     }
